misys_bridge_onprem/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-compress>=1.13
waitress>=2.1.0
pyodbc>=4.0.0
pymssql>=2.2.0
python-dotenv>=1.0.0
